fix: realizado alterações para que o sistema funcione no celular; fix: removido bibliotecas e alterado conexao com arquivo do banco de dados;
