+++ GerenciaAgro/MauiProgram.cs
@@ -7,8 +7,6 @@ using CasosDeUso.Interface.InterfaceCultivoUseCase;
 using CasosDeUso.Interface.InterfacePragaUseCase;
 using CasosDeUso.PluginsInterfaces;
 using CasosDeUso.PragaCasoDeUso;
-using CoreBusiness.Entidades;
-using DadosEmMemoria;
 using GerenciaAgro.Views;
 using GerenciaAgro.Views.Controls;
 using Microsoft.Extensions.Logging;
@@ -23,99 +21,54 @@ namespace GerenciaAgro
         {
             try
             {
+                var builder = MauiApp.CreateBuilder();
 
-            var builder = MauiApp.CreateBuilder();
+                var culture = new CultureInfo("pt-BR");
+                CultureInfo.DefaultThreadCurrentCulture = culture;
+                CultureInfo.DefaultThreadCurrentUICulture = culture;
 
-            var culture = new CultureInfo("pt-BR");
-            CultureInfo.DefaultThreadCurrentCulture = culture;
-            CultureInfo.DefaultThreadCurrentUICulture = culture;
-
-            builder
-                .UseMauiApp<App>()
-                .ConfigureFonts(fonts =>
-                {
-                    fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
-                    fonts.AddFont("OpenSans-Semibold.ttf", "OpenSansSemibold");
-                });
+                builder
+                    .UseMauiApp<App>()
+                    .ConfigureFonts(fonts =>
+                    {
+                        fonts.AddFont("OpenSans-Regular.ttf", "OpenSansRegular");
+                        fonts.AddFont("OpenSans-Semibold.ttf", "OpenSansSemibold");
+                    });
 
 #if DEBUG
-            builder.Logging.AddDebug();
+                builder.Logging.AddDebug();
 #endif
-            #region injeção de dependências
-            builder.Services.AddSingleton<IRepositorioDeAplicacao, RepositorioAplicacaoSqlLite>();
-            builder.Services.AddSingleton<IRepositorioDeAgrotoxico, RepositorioAgrotoxicoSqlLite>();
-            builder.Services.AddSingleton<IRepositorioDeCultivo, RepositorioCultivoSqlLite>();
-            builder.Services.AddSingleton<IRepositorioDePraga, RepositorioPragaSqlLite>();
+                #region injeção de dependências
+                builder.Services.AddSingleton<IRepositorioDeAplicacao, RepositorioAplicacaoSqlLite>();
 
-            //builder.Services.AddSingleton<IRepositorioDeAplicacao, Dados>();
-            builder.Services.AddSingleton<IVisualizarAplicacaoUseCase, VisualizarAplicacaoUseCase>();
-            builder.Services.AddSingleton<IApagarAplicacaoUseCase, ApagarAplicacaoUseCase>();
-            builder.Services.AddSingleton<IAdicionarAplicacaoUseCase, AdicionarAplicacaoUseCase>();
-            builder.Services.AddSingleton<IEditarAplicacaoUseCase, EditarAplicacaoUseCase>();
-            builder.Services.AddSingleton<IVisualizarCultivoUseCase, VisualizarCultivoUseCase>();
-            builder.Services.AddSingleton<IVisualizarAgrotoxicoUseCase, VisualizarAgrotoxicoUseCase>();
-            builder.Services.AddSingleton<IVisualizarPragaUseCase, VisualizarPragaUseCase>();
-            builder.Services.AddSingleton<IAdicionarAgrotoxicoUseCase, AdicionarAgrotoxicoUseCase>();
-            builder.Services.AddSingleton<IAdicionarPragaUseCase, AdicionarPragaUseCase>();
-            builder.Services.AddSingleton<IAdicionarCultivoUseCase, AdicionarCultivoUseCase>();
-            builder.Services.AddSingleton<IEditarAplicacaoUseCase, EditarAplicacaoUseCase>();
-            builder.Services.AddSingleton<IApagarAplicacaoUseCase, ApagarAplicacaoUseCase>();
-            builder.Services.AddTransient<AplicacaoControle>();
-            #endregion
-            builder.Services.AddSingleton<AplicacaoPagina>();
-            builder.Services.AddSingleton<SelecaoItemCadastro>();
-            builder.Services.AddSingleton<AplicacaoControle>();
-            builder.Services.AddSingleton<AgrotoxicoControle>();
-            builder.Services.AddSingleton<PragaControle>();
-            builder.Services.AddSingleton<CultivoControle>();
-            builder.Services.AddSingleton<AgrotoxicoLista>();
-            builder.Services.AddSingleton<AplicacaoLista>();
+                builder.Services.AddSingleton<IVisualizarAplicacaoUseCase, VisualizarAplicacaoUseCase>();
+                builder.Services.AddSingleton<IApagarAplicacaoUseCase, ApagarAplicacaoUseCase>();
+                builder.Services.AddSingleton<IAdicionarAplicacaoUseCase, AdicionarAplicacaoUseCase>();
+                builder.Services.AddSingleton<IEditarAplicacaoUseCase, EditarAplicacaoUseCase>();
+                builder.Services.AddSingleton<IVisualizarCultivoUseCase, VisualizarCultivoUseCase>();
+                builder.Services.AddSingleton<IVisualizarAgrotoxicoUseCase, VisualizarAgrotoxicoUseCase>();
+                builder.Services.AddSingleton<IVisualizarPragaUseCase, VisualizarPragaUseCase>();
+                builder.Services.AddSingleton<IAdicionarAgrotoxicoUseCase, AdicionarAgrotoxicoUseCase>();
+                builder.Services.AddSingleton<IAdicionarPragaUseCase, AdicionarPragaUseCase>();
+                builder.Services.AddSingleton<IAdicionarCultivoUseCase, AdicionarCultivoUseCase>();
+                builder.Services.AddTransient<AplicacaoControle>();
+                #endregion
+                builder.Services.AddSingleton<AplicacaoPagina>();
+                builder.Services.AddSingleton<SelecaoItemCadastro>();
+                builder.Services.AddSingleton<AplicacaoControle>();
+                builder.Services.AddSingleton<AgrotoxicoControle>();
+                builder.Services.AddSingleton<PragaControle>();
+                builder.Services.AddSingleton<CultivoControle>();
+                builder.Services.AddSingleton<AgrotoxicoLista>();
+                builder.Services.AddSingleton<AplicacaoLista>();
 
-            //builder.Services.AddSingleton<EditarContatoPage>();
-            //builder.Services.AddSingleton<AdicionarContatoPage>();
-
-            var app = builder.Build();
-
-            SeedDatabase(app.Services);
-
-            return app;
-            }catch (Exception ex)
+                return builder.Build();
+            }
+            catch (Exception ex)
             {
                 Console.WriteLine($"Erro ao criar o aplicativo: {ex.Message}");
-                throw; 
+                throw;
             }
         }
-
-        private static async void SeedDatabase(IServiceProvider services)
-        {
-            var cultivoRepo = services.GetRequiredService<IRepositorioDeCultivo>();
-            var pragaRepo = services.GetRequiredService<IRepositorioDePraga>();
-            var agrotoxicoRepo = services.GetRequiredService<IRepositorioDeAgrotoxico>();
-
-            // Exemplo: Cultivos
-            var cultivos = await cultivoRepo.BuscarCultivo("");
-            if (cultivos.Count == 0)
-            {
-                await cultivoRepo.AdicionarCultivoAsync(new Cultivo { Nome = "Soja", Id = Guid.NewGuid() });
-                await cultivoRepo.AdicionarCultivoAsync(new Cultivo { Nome = "Milho", Id = Guid.NewGuid() });
-            }
-
-            // Exemplo: Pragas
-            var pragas = await pragaRepo.BuscarPraga("");
-            if (pragas.Count == 0)
-            {
-                await pragaRepo.AdicionarPragaAsync(new Praga { Nome = "Lagarta", Id = Guid.NewGuid() });
-                await pragaRepo.AdicionarPragaAsync(new Praga { Nome = "Pulgão", Id = Guid.NewGuid() });
-            }
-
-            // Exemplo: Agrotóxicos
-            var agrotoxicos = await agrotoxicoRepo.BuscarAgrotoxico("");
-            if (agrotoxicos.Count == 0)
-            {
-                await agrotoxicoRepo.AdicionarAgrotoxico(new Agrotoxico { Nome = "Inseticida X", Lote = "XYYWH2509", Validade = DateTimeOffset.Now.AddMonths(2), Id = Guid.NewGuid() });
-                await agrotoxicoRepo.AdicionarAgrotoxico(new Agrotoxico { Nome = "Fungicida Y", Lote = "POIII322", Validade = DateTimeOffset.Now.AddMonths(6), Id = Guid.NewGuid() });
-            }
-        }
-
     }
 }
