+++ SqlLite/SqlLite.csproj
@@ -24,11 +24,6 @@
 		<PackageReference Include="PdfSharpCore" Version="1.3.67" />
 		<PackageReference Include="sqlite-net-pcl" Version="1.9.172" />
 		<PackageReference Include="SQLitePCLRaw.bundle_green" Version="2.1.11" />
-		<PackageReference Include="SQLitePCLRaw.core" Version="2.1.11" />
-		<PackageReference Include="SQLitePCLRaw.lib.e_sqlite3" Version="2.1.11" />
-		<PackageReference Include="SQLitePCLRaw.lib.e_sqlite3.android" Version="2.1.11" />
-		<PackageReference Include="SQLitePCLRaw.provider.e_sqlite3" Version="2.1.11" />
-		<PackageReference Include="SQLitePCLRaw.provider.sqlite3" Version="2.1.11" />
 	</ItemGroup>
 
 	<ItemGroup>
